refactor(repository): merge FindPublishedPage/FindAdminPage into findCatalogPage and split master_catalog_read.go
